requirements.txt
setuptools~=57.0.0
numpy~=1.22.4
scikit-learn~=1.1.2
hashids~=1.3.1
PyYAML~=6.0
dotmap~=1.3.30